reduce scope of changes
